+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Enhancement: Added 'max_freebusy_occurrences` setting to avoid potential DOS on reports
 * Enhancement: remove unexpected control codes from uploaded items
 * Enhancement: add 'strip_domain' setting for username handling
+* Enhancement: add option to toggle debug log of right with doesn't match
 * Drop: remove unused requirement "typeguard"
 * Improve: Refactored some date parsing code